AI search Agent/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-tavily>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.40.0
